...Actually put content in `Home`, with pointless snackbar/FAB
Oops I forgot to commit `Home`. Not going to mess with git or
try to even amend previous message.

Floating action button to open snackbar just for the sake of it.
Opacity 50% because I don't feel like figuring out how to make it
show up properly without that hack at the moment.