Render JSX/HTML from GraphCMS in `GraphQLDemo FC` &` CodeDialog FC`
Ok, this is a lot of stuff... sorry. I got obsessed with getting
all of this working from the importing of queried data to rendering
as highlighted JSX. Explanation:

`GraphQLDemo FC` now pulls two queries to display content. The
'h2/p1/p2' from 'graphQLDemo' as it already exists now exists
alongside another query, 'demoCode' that grabs
'title/fileName/gitHubLink/richText (HTML)/multiline' from a different
type of content entry. The demo page currently has `demoCode.multiline`
embedded in it through the `<SyntaxHighlighter` library component.
The highlighter can turn strings, such as 'multiline', and show them
as `<code>` blocks highlighted like the respective language would be
in an IDE/text editor.

So basically, I put the code of the `GraphQLDemo FC` component
into GraphQL as a string and it shows up on the page looking like
a text editor's view of the code. However, this code preview will be
moved to a new component: `CodeDialog FC`.

`CodeDialog FC` opens up when a "View Code" button is clicked.
It shows the title of the code, file name, permalink to file on GH,
and a very crude rendering of the HTML of the code from the page.
The JSX highlighting from `GraphQLDemo FC` will be moved out of there
and placed into `CodeDialog FC`. I am commiting it as is right now to
highlight some of the bigger ideas that I learned along the way...

1) React renders HTML through its implementation of `innerHTML`, which
it calls `dangerouslySetInnerHTML`. This is to note that injecting
HTML into the DOM can open up the site to vulnerabilities like XSS.
To circumvent this, I use the "DomPurity" library to sanitize it.
There was alternatives that rolled up the injecting and sanitizing
into one component to inject things, but I prefer this approach
because it highlights the dangers of injecting HTML directly into
the DOM.

2) Export formats of rich text can be tricky. GraphQL offers
rich text as HTML/Markdown/Raw/text. Raw is it's own thing I won't
get into, but HTML/MD/txt all inject their own `\n` and `\t` syntax
to represent newlines and tabs, respectively. Injecting HTML into the
DOM naturally called for HTML, but then I couldn't use highlighting
libraries. The multiple libraries that I tried wanted only strings.
But they were not straight forward in their docs on how to format
the text such that the newlines and tabs are rendered rather than just
their special characters in text. So I spent a lot of time trying to
figure out how to get the code to render properly. Ultimately,
none of the options on rich text worked perfectly, and I realize that
rich text was overkill. GraphCMS offered multiline text, and it
*just worked*. No overhead from rich text, even its own 'text' type
obstructed the highlighter's input and output process.

In retrospect, this doesn't look so complicated and I was overthinking
certain aspects and not thinking of more basic solutions. But in the
end I learned about displaying various data formats safely in the
DOM, so that's a good take away.

The next step will be to move the JSX block w/highlighting from the
demo component to the dialog component, and making the code toggle
component function. But I have to go to work soon, and this is a
huge commit anyways.