Add devtools config for Apollo Client, remove a query of field
Apollo Client has a devtools plugin with a lot of cool features.
Chrome link: https://chrome.google.com/webstore/detail/apollo-client-developer-t/jdkknkkbebbapilgoeccciglkfbmbnfm

When I enabled the the plugin with that setting, I was able to run
queries in a dedicated "Apollo" tab in the devtools. This helped
remind me that I was still querying an HTML field that I no longer
used in `CodeDialog FC`.