Toggle showcase code
`Showcase FC` was converted to `Showcase CC` because it needed a
boolean `showCode: false` to default to. The `toggleCode` function
changes that state, and `Togglecode FC` has a click handler to respond.
`GraphQLDemo FC` then conditionally shows the icons to show code
and the "View Code" button if code is enabled.