Add photo of component hierarchy to `GraphQLDemo FC`
Proud of my photoshop. Photo is from the React Developer Tools
plugin's tab for showing component hierarchy.

Also I only looked at this on mobile and it looked fine, but it is
BLOWN UP and HUGE on desktop. Fine for now, going to bed. Proud of
the stuff I have done today in the few hours I have before and
after work.