Fix body to be size of viewport
This was making me insane and has happened in different projects,
React or not. Basically, the `index.js` and `app.js` and `index.html`
all have some sort of say in how React is injected into the DOM and/ or
presented.

I had this bandaid "fix" where the body had the background of the
React app background, but this didn't solve the root problem. Imagine
that the body was 500px. The React portion of a page, like "Home",
would maybe be like 200px or 300px at the top and the rest of the body
below it was empty beside from the background color.

Putting in this change to max the height of `<html>`, `<body>`, `#root`,
and `.App` in index.css fixes this.

Also, body gets the background color for good measure.