Hide section code toggle and pad button on `RecruiterView FC`
I thought I had given padding to the button that was hugging the
bottom of the page to show the code.

Anyways, this also hides each individual code toggling of a section.
Recruiters may not want to see all of the code, but toggle it for
a section.

Addresses #9 (Create recruiter's showcase view)
Project: 'Recruiter View'