Add commented out `ContactPlatform` component, please read reasoning
I include this if it is more relevant in the future and as a learning
note because that's what this portfolio project is all about.

The format of the contact page's `<h3>{{platform}}</h3> <a href />`
seemed to ask for a component in itself. However, to be sufficiently
generic I thought that it should be able to be passed references to
MUI icons. However, I couldn't do some in a naive way as just a
platform string that conditionally renders an icon. Furthermore,
being able to handle importing whatever icon is relavant would be...
interesting but likely a tricky.

Furthermore, `<Obfuscate />` has a few different attributes. I use
email in particular. Checking for an optional prop in this component
was fine, but I needed to pass what type of prop to set as the attribute
as, and that didn't seem straightforward, at least from how I was
searching for guides.

All in all something could likely be done and cool to figure out,
but for now would be quite overkill.