Create `CopyToClipB FC` that copies input to clipboard
Shortened the name to `ClipB` instead of `Clipboard` because I was
having namespace issues with the package. Technically I should be
able to use the alias of the import as its own space, but I thought
that this component name distinction would be more clear.

Vanilla HTML forms.

I present this on the `CompWorkshop` page at `/componentWorkshop URL`,
which I created to look at components I am developing in a vacuum.

Addresses #47 (Add copy-to-clipboard functionality)
Project: 'Easy CV Info'