Comment out CompWorkshop to test deployment
Of note, for some reason VSC doesn't recognize `CompWorkshop FC` as
a function. I must be missing something major, but it looks like
just like I should be doing it. Perhaps the folder being `/utils`
means it is a reserved sort of folder, or because it isn't capitalized?

VSC does give `/utils` a special folder icon. Will look into it,
but I am about to push this up and deploy on staging.

Addresses #51 (Fix memory leak on deployment)