Group showcase demos and views into `/Demos` and `/RoleViews` folders
`/Demos` has only one demo, the GraphQL one. But that will be good
when more come. Similarly, only the recruiter view is in `/RoleViews`.

`ViewPreview FC` is still in the root of `/Showcase`, as it is a
utility of `Showcase CC` and not a view itself.

Addresses #52 (Restructure `/showcase` component hierarchy)
Projects: 'GraphQL Showcase' and `Showcase View Selection`